some changes to code generation
